Added a list of the animals
this is a test version. and some java script will be adjusted soon after,

The short summary is: small images of the animals,  first 200 characters of the description, what the animal eats,  a link to the group page (reptile, bird, mammal).                         - sidebar animal clicked = short summary of that animal you clicked (the summary should replace the wellcome message) .                                      - sidebar animal clicked = short summary of that animal you clicked .(the summary should replace the welcome message)                                     - if the "active" animal is clicked, it should not be active anymore and welcome message is desplayed